chore: moved msg definition to `project-name/msg.py`
