+++ projects/owa-env-desktop/owa_env_desktop/keyboard_mouse/__init__.py
@@ -3,8 +3,3 @@ from . import callables  # noqa
 
 # Register listeners
 from . import listeners  # noqa
-
-from .msg import KeyboardEvent, MouseEvent
-
-
-__all__ = ["KeyboardEvent", "MouseEvent"]
