+++ projects/owa-env-desktop/owa_env_desktop/window/__init__.py
@@ -1,6 +1,2 @@
 # Register callables
 from . import callables  # noqa
-from .msg import WindowInfo
-
-
-__all__ = ["WindowInfo"]
